requirements.txt
Flask==2.2.5
matplotlib==3.4.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
Pillow==10.2.0
Requests==2.31.0
scikit_learn==1.0.2
tensorflow==2.11.0
tensorflow==2.4.0
tensorflow_datasets==4.8.2
tensorflow_intel==2.11.0
